Refactor ProtectedRouteProps type definition; move to new types directory and remove old file
